+++ Assets/LDtkUnity/Runtime/Data/Extensions/Level/NeighbourLevel.cs
@@ -5,17 +5,16 @@ namespace LDtkUnity
     public partial class NeighbourLevel
     {
         //Using an implicit operator for simpler direct access to the neighbour's level data. (May change if problematic)
+        public static implicit operator Level(NeighbourLevel neighbour) => neighbour.Level;
+        
         /// <summary>
         /// Reference to the level of this neighbour. 
         /// </summary>
-        /// <param name="neighbour">
-        /// The neighbour.
-        /// </param>
         /// <returns>
         /// The level of this neighbour. <br/>
         /// Make sure to call <see cref="LDtkIidBank"/>.<see cref="LDtkIidBank.CacheIidData"/> first!
         /// </returns>
-        public static implicit operator Level(NeighbourLevel neighbour) => neighbour.LevelIid == null ? null : LDtkIidBank.GetIidData<Level>(neighbour.LevelIid);
+        public Level Level => LevelIid == null ? null : LDtkIidBank.GetIidData<Level>(LevelIid);
 
         /// <value>
         /// Returns true if this neighbour is above the relative level.