+++ Assets/LDtkUnity/Runtime/Fields/LDtkFieldElement.cs
@@ -268,5 +268,26 @@ namespace LDtkUnity
                     return false;
             }
         }
+        
+        //special piece for neighbours. the int value is the char for direction
+        public LDtkFieldElement(NeighbourLevel neighbour)
+        {
+            _canBeNull = false;
+            _isNotNull = true;
+            _type = LDtkFieldType.EntityRef;
+            _string = neighbour.LevelIid;
+            _int = neighbour.Dir[0];
+        }
+
+        public NeighbourLevel AsNeighbourLevel()
+        {
+            string[] elements = _string.Split('_');
+            NeighbourLevel neighbour = new NeighbourLevel
+            {
+                Dir = $"{(char)_int}",
+                LevelIid = _string
+            };
+            return neighbour;
+        }
     }
 }